+++ src/main/kotlin/gg/match/domain/board/overwatch/repository/HeroRepository.kt
@@ -5,6 +5,6 @@ import gg.match.domain.board.overwatch.entity.Type
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface HeroRepository: JpaRepository<Hero, Long> {
-    fun findByNameAndBattletagAndType(name: String, battletag: Int, type: Type): Hero
+    fun findByNameAndBattletagAndType(name: String, battletag: Long, type: Type): Hero
     fun findByName(name: String): Hero
 }
